% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenprob.R
\name{betweenprob}
\alias{betweenprob}
\title{The in between probability function}
\usage{
betweenprob(x, y, n, p)
}
\arguments{
\item{x}{a number, which represents the upper bound of the probability to be found.}

\item{y}{a number, which represents the lower bound of the probability to be found.}

\item{n}{a number, the number of independent trials or experiments to be represented.}

\item{p}{a decimal number, represents the probability of success on a single trial.}
}
\value{
a number
}
\description{
The in between probability function
}
\examples{
betweenprob(x = 2,y = 8,n = 10,p = 0.6)
}
