% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{bootpval function}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{the sample data}

\item{conf.level}{the confidence level}

\item{iter}{how many iterations}

\item{mu0}{the mean to be tested}

\item{test}{the type of test to be used}
}
\description{
bootpval function
}
