% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM Data}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Description of column1}
  \item{DAMAGE}{Description of column2}
  ...
}
}
\source{
<provide relevant source or reference>
}
\usage{
fire
}
\description{
Data on fire damages collected from various sources.
}
\keyword{datasets}
