% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations. Default is 10000.}

\item{x}{A numeric vector of data values to be bootstrapped.}

\item{fun}{A function to compute the statistic of interest. Default is "mean".}

\item{alpha}{The significance level for the confidence interval. Default is 0.05.}

\item{cx}{Scaling factor for text size in plots. Default is 1.5.}

\item{...}{Additional parameters passed to the histogram plotting function.}
}
\value{
A list containing the confidence interval, the function used, the original data, and the bootstrap statistics.
}
\description{
This function performs bootstrap resampling on the provided numeric vector `x`
and calculates the specified statistic using the specified function (`fun`).
It also generates a histogram of the bootstrap statistics along with confidence intervals.
}
\examples{
set.seed(123)
results <- myboot2(iter = 10000, x = rnorm(100), fun = mean, alpha = 0.05, cx = 1.5)
}
