% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Estimation function}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a vector representing the sample data}

\item{mu}{vector value of the mean}

\item{sig}{A vector value of the standard deviation}

\item{...}{Additional parameter}
}
\value{
a graphic and x/y coords to a point
}
\description{
The function estimates the maximum likelihood for the mean and standard deviation of a normal distribution based on a given vector
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
