% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{The plot function}
\usage{
myplot(x, y)
}
\arguments{
\item{x}{a vector of quantitative data}

\item{y}{a vector of quantitative data}
}
\value{
a plot
}
\description{
The plot function
}
\examples{
myplot(x = 1.10, y = 2.5)

}
