% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{The square function}
\usage{
mysquare(x)
}
\arguments{
\item{x}{a vector of quantitative data}
}
\value{
a vector
}
\description{
The square function
}
\examples{
mysquare(x = 1.10)
}
