% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats on the plane}

\item{gamma}{Probability that more passengers show up than there are seats}

\item{p}{Probability that a ticketed passenger will show up for the flight}
}
\value{
blah
}
\description{
ntickets
}
\examples{
ntickets(N = 100, gamma = 0.05, p = 0.9)
}
